backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
motor==3.4.0
pymongo==4.6.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
pydantic[email]==2.8.2
pydantic-settings==2.4.0
python-dateutil==2.9.0.post0
requests==2.32.3
google-generativeai==0.7.2
